huggingface_Gradio/requirements.txt
gradio==4.8.0
numpy==1.26.2
opencv_python==4.7.0.72
Pillow==10.1.0
sahi==0.11.15
torch==2.1.1
ultralytics==8.0.223
